requirements.txt
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.5
cryptography==3.4.7
Flask==1.1.2
flask-migrate==3.0.0
flask-sqlalchemy==2.5.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dotenv==0.17.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.10.0
urllib3==1.26.4
Werkzeug==1.0.1
